\name{morseStan-package}
\alias{morseStan-package}
\alias{morseStan}
\docType{package}
\title{
\packageTitle{morseStan}
}
\description{
\packageDescription{morseStan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{morseStan}
\packageIndices{morseStan}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{morseStan}

Maintainer: \packageMaintainer{morseStan}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
